+++ addon/components/flexberry-lookup.js
@@ -5,6 +5,7 @@
 import { assert, warn, debug } from '@ember/debug';
 import { inject as service } from '@ember/service';
 import { computed, observer, get } from '@ember/object';
+import { deprecatingAlias } from '@ember/object/computed';
 import { later, run, once } from '@ember/runloop';
 import { merge } from '@ember/polyfills';
 import { isNone } from '@ember/utils';
@@ -91,6 +92,28 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
     @type Boolean
   */
   _cachedDropdownValue: undefined,
+
+  /**
+    @private
+    @property _pageInResultsForAutocomplete
+    @type Number
+    @default 1
+  */
+  _pageInResultsForAutocomplete: 1,
+
+  /**
+    Computed property for {{#crossLink "FlexberryLookup/modalDialogSettings:property"}}modalDialogSettings{{/crossLink}} property.
+
+    @private
+    @property _modalDialogSettings
+  */
+  _modalDialogSettings: computed('modalDialogSettings', function () {
+    return merge({
+      sizeClass: this.get('_sizeClass'),
+      useOkButton: false,
+      useCloseButton: false,
+    }, this.get('modalDialogSettings'));
+  }),
 
   /**
     Text to be displayed in field, if value not selected.
@@ -196,6 +219,15 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
     @default false
   */
   autocomplete: false,
+
+  /**
+    If `true`, page switching buttons will be available in the results for autocomplete.
+
+    @property usePaginationForAutocomplete
+    @type Boolean
+    @default false
+  */
+  usePaginationForAutocomplete: false,
 
   /**
     Flag to show that lookup is in dropdown mode.
@@ -422,6 +454,32 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
   lookupAdditionalLimitFunction: undefined,
 
   /**
+    An object with options and settings for the modal dialog of the component.
+
+    The following options and settings are available:
+      - {{#crossLink "ModalDialog/sizeClass:property"}}`sizeClass`{{/crossLink}}, not set by default.
+      - {{#crossLink "ModalDialog/useOkButton:property"}}`useOkButton`{{/crossLink}}, default `false`.
+      - {{#crossLink "ModalDialog/useCloseButton:property"}}`useCloseButton`{{/crossLink}}, default `false`.
+      - {{#crossLink "ModalDialog/settings:property"}}`settings`{{/crossLink}}, not set by default.
+
+    @property modalDialogSettings
+    @type Object
+  */
+  modalDialogSettings: undefined,
+
+  /**
+    Sets the {{#crossLink "ModalDialog/sizeClass:property"}}`sizeClass`{{/crossLink}} property for the modal dialog of the component.
+
+    @property sizeClass
+    @type String
+    @deprecated Use `modalDialogSettings.sizeClass`.
+  */
+  sizeClass: deprecatingAlias('_sizeClass', {
+    id: 'flexberry-lookup.modal-dialog-size-class',
+    until: '3.0',
+  }),
+
+  /**
     This computed property forms a set of properties to send to lookup window.
     Closure action `lookupWindowCustomProperties` is called here if defined,
     otherwise `undefined` is returned.
@@ -501,6 +559,7 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
     'title',
     'lookupLimitPredicate',
     'relatedModel',
+    '_modalDialogSettings',
     '_lookupWindowCustomPropertiesData',
     'inHierarchicalMode',
     'hierarchicalAttribute',
@@ -519,9 +578,7 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
         folvComponentName: this.get('folvComponentName'),
         inHierarchicalMode: this.get('inHierarchicalMode'),
         hierarchicalAttribute: this.get('hierarchicalAttribute'),
-
-        //TODO: move to modal settings.
-        sizeClass: this.get('sizeClass')
+        modalDialogSettings: this.get('_modalDialogSettings'),
       };
     }),
 
@@ -819,7 +876,8 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
         showInSeparateRoute: this.get('previewOnSeparateRoute'),
         modelName: relationModelName,
         controller: this.get('controllerForPreview'),
-        projection: this.get('previewFormProjection')
+        projection: this.get('previewFormProjection'),
+        modalDialogSettings: this.get('_modalDialogSettings'),
       };
 
       /* eslint-disable ember/closure-actions */
@@ -854,6 +912,15 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
     Observe lookup value changes.
   */
   valueObserver: observer('value', 'displayAttributeName', function() { once(this, '_valueObserver'); }),
+
+  /**
+    Observe lookup autocomplete max results changes.
+  */
+  maxResultsObserver: observer('maxResults', function() {
+    if (this.get('autocomplete')) {
+      this._onAutocomplete();
+    }
+  }),
 
   /**
     It seems to me a mistake here, it should be [willDestroyElement](http://emberjs.com/api/classes/Ember.Component.html#event_willDestroyElement).
@@ -935,7 +1002,7 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
     @private
   */
   _setModalIsStartToShow(componentName) {
-    if (this.get('componentName') === componentName) {
+    if (this.get('componentName') === componentName && this.get('modalIsShow')) {
       this.set('modalIsBeforeToShow', false);
       this.set('modalIsStartToShow', true);
     }
@@ -1020,7 +1087,9 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
     let i18n = _this.get('i18n');
     this.$().search({
       minCharacters: minCharacters,
-      maxResults: maxResults + 1,
+
+      // +2 for page switch buttons.
+      maxResults: maxResults + 2,
       cache: false,
       templates: {
         /* eslint-disable no-unused-vars */
@@ -1057,34 +1126,26 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
             builder.where(resultPredicate);
           }
 
-          let maxRes = _this.get('maxResults');
-          let iCount = 1;
-          builder.top(maxRes + 1);
+          const usePagination = _this.get('usePaginationForAutocomplete');
+          const skip = maxResults * (_this.get('_pageInResultsForAutocomplete') - 1);
+          builder.skip(skip);
+          builder.top(maxResults);
           builder.count();
 
           run(() => {
             store.query(relationModelName, builder.build()).then((records) => {
-              callback({
-                success: true,
-                results: records.map(i => {
-                  let attributeName = i.get(displayAttributeName);
-                  if (iCount > maxRes && records.meta.count > maxRes) {
-                    return {
-                      title: '...',
-                      noResult: true
-                    };
-                  } else {
-                    iCount += 1;
-                    return {
-                      title: attributeName,
-                      instance: i
-                    };
-                  }
-                })
-              });
-              run.next(() => {
-                _this.onShowHide(true);
-              });
+              const results = records.map((r) => ({ title: r.get(displayAttributeName), instance: r }));
+
+              if (usePagination && skip > 0) {
+                results.unshift({ title: '<div class="ui center aligned container"><i class="angle up icon"></i></div>', prevPage: true });
+              }
+
+              if (skip + records.get('length') < records.get('meta.count')) {
+                const title = usePagination ? '<div class="ui center aligned container"><i class="angle down icon"></i></div>' : '...';
+                results.push({ title: title, nextPage: usePagination });
+              }
+
+              callback({ success: true, results: results });
             }, () => {
               callback({ success: false });
             });
@@ -1115,8 +1176,9 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
        * @param {Object} result Item from array of objects, built in `responseAsync`.
        */
       onSelect(result) {
-        if (!result.noResult) {
+        if (result.instance) {
           state = 'selected';
+          _this.set('_pageInResultsForAutocomplete', 1);
 
           run(() => {
             debug(`Flexberry Lookup::autocomplete state = ${state}; result = ${result}`);
@@ -1129,6 +1191,24 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
                 newRelationValue: result.instance
               });
           });
+        } else if (_this.get('usePaginationForAutocomplete')) {
+          state = 'loading';
+          if (result.nextPage) {
+            _this.incrementProperty('_pageInResultsForAutocomplete');
+          }
+
+          if (result.prevPage) {
+            _this.decrementProperty('_pageInResultsForAutocomplete');
+          }
+
+          // In the `Semantic UI` of version 2.1.7, the results are closed regardless of what the `onSelect` handler returns.
+          // This function allows us to start the search again, thanks to the `searchOnFocus` option.
+          later(() => {
+            _this.$().search('add results', '');
+            _this.$('input').focus();
+          }, 500);
+
+          return false;
         }
       },
 
@@ -1140,7 +1220,8 @@ export default FlexberryBaseComponent.extend(FixableComponent, {
         // Set displayValue directly because value hasn'been changes
         // and Ember won't change computed property.
 
-        if (state !== 'selected') {
+        if (state !== 'selected' && state !== 'loading') {
+          _this.set('_pageInResultsForAutocomplete', 1);
           let displayValue = _this.get('displayValue');
           if (displayValue) {
             if (_this.get('autocompletePersistValue')) {